refactor(execution): modularize backends
